requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
twilio==9.2.3
requests==2.32.3
faster-whisper==1.0.3
google-generativeai==0.8.2
upstash-redis==1.2.0
edge-tts==6.1.12
python-multipart==0.0.12
aiofiles==24.1.0

